ai_engine/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf
